+++ packages/harpstrata/src/harpstrata/IsActive/getActiveIdsPair/utils/getCounterpartActiveIds/index.ts
